requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-summernote==0.8.11.1
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
djangorestframework-jwt==1.11.0
djoser==1.3.2
inflection==0.3.1
npm==0.1.1
optional-django==0.1.0
psycopg2==2.7.6.1
PyJWT==1.7.1
pytz==2018.7
six==1.12.0
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
vue==0.0.1
